# openapi.yaml
openapi: 3.0.3
info:
  title: Whop SaaS Marketplace API
  description: |
    Production-grade marketplace API enabling multi-tenant SaaS, Stripe Connect payments,
    license management, and webhook event broadcasting.

    ## Authentication
    All endpoints require authentication via session token cookie (`session_token`).

    ## Multi-Tenancy
    All resources are scoped to companies. Users can access multiple companies with different roles.

    ## Rate Limiting
    - 100 requests per minute per IP
    - Returns `429 Too Many Requests` when exceeded

    ## Security
    - Tenant isolation via `company_id` filtering
    - Role-based access control (owner/admin/member)
    - Money stored in minor units (cents) to avoid floating-point errors
  version: 1.0.0
  contact:
    name: API Support
    email: api@whopsaas.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.whopsaas.com
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: JWT session token (HTTP-only cookie)
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod_abc123
        company_id:
          type: string
          example: biz_xyz789
        title:
          type: string
          example: Pro Plan
        description:
          type: string
          nullable: true
          example: Premium features with priority support
        image_url:
          type: string
          format: uri
          nullable: true
        price_amount:
          type: number
          format: decimal
          example: 49.99
        price_minor_units:
          type: integer
          example: 4999
          description: Price in cents (avoid floating-point errors)
        currency:
          type: string
          example: USD
        plan_type:
          type: string
          enum: [one_time, monthly, yearly]
        billing_period:
          type: integer
          nullable: true
          description: Billing period in days (30, 365, etc.)
        trial_days:
          type: integer
          example: 14
        active:
          type: boolean
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - company_id
        - title
        - price_amount
        - price_minor_units
        - currency
        - plan_type
    ProductCreate:
      type: object
      properties:
        company_id:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        image_url:
          type: string
          format: uri
        price_amount:
          type: number
          description: Price in dollars (e.g., 49.99)
        price_minor_units:
          type: integer
          description: Price in cents (e.g., 4999) - provide EITHER this OR price_amount
        currency:
          type: string
          default: USD
        plan_type:
          type: string
          enum: [one_time, monthly, yearly]
        billing_period:
          type: integer
        trial_days:
          type: integer
          default: 0
        active:
          type: boolean
          default: true
        metadata:
          type: object
      required:
        - company_id
        - title
        - plan_type
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          example: whk_abc123
        company_id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - payment.succeeded
              - payment.failed
              - membership.created
              - membership.updated
              - membership.canceled
              - product.created
              - product.updated
              - product.deleted
        api_version:
          type: string
          default: v1
        secret:
          type: string
          description: HMAC secret (masked after creation)
          example: whsec_abc123...
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            deliveries:
              type: integer
...
    WebhookUpdate:
      type: object
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          minItems: 1
        api_version:
          type: string
        active:
          type: boolean

    # ========================================================================
    # Common Schemas
    # ========================================================================
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
security:
  - cookieAuth: []
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: List products
      description: List all products for a company with pagination
      parameters:
        - name: company_id
          in: query
          required: true
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad request (missing company_id)
        '403':
          description: Forbidden (no access to company)

    post:
      tags:
        - Products
      summary: Create product
      description: Create a new product (requires admin/owner role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
        '403':
          description: Forbidden (insufficient permissions)
...
  /api/webhooks/{id}:
    get:
      tags:
        - Webhooks
      summary: Get webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'

    put:
      tags:
        - Webhooks
      summary: Update webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'

    delete:
      tags:
        - Webhooks
      summary: Delete webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string